requirements.txt
numpy>=1.24.0
scipy>=1.9.0

# Qiskit and Quantum Kernel functionality
qiskit-terra==0.25.1
qiskit-aer==0.12.2
qiskit-machine-learning==0.7.1

# Development & testing
pytest>=7.0.0
pytest-cov>=4.0.0
pigar>=0.10.0
reservoirpy[hyper]>=0.3.13
hyperopt
matplotlib
jsonargparse